data-migration-etl/requirements.txt
mysql-connector-python
psycopg2-binary